requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\andre\Documents\Barney Bot\barney.py: 28
# C:\Users\andre\Documents\Barney Bot\qr_test.py: 5
Pillow == 8.3.1

# C:\Users\andre\Documents\Barney Bot\barney.py: 25
# C:\Users\andre\Documents\Barney Bot\schoology_test.py: 2
PyYAML == 5.4.1

# C:\Users\andre\Documents\Barney Bot\barney.py: 32
confusables == 1.2.0

# C:\Users\andre\Documents\Barney Bot\barney.py: 25
# C:\Users\andre\Documents\Barney Bot\schoology_test.py: 2
cronyo == 0.4.5

# C:\Users\andre\Documents\Barney Bot\barney.py: 21,26,27
# C:\Users\andre\Documents\Barney Bot\barney_backup.py: 3,4
# C:\Users\andre\Documents\Barney Bot\barney_chess.py: 1
# C:\Users\andre\Documents\Barney Bot\barney_mafia_test.py: 3,4
# C:\Users\andre\Documents\Barney Bot\barney_minimal.py: 4
# C:\Users\andre\Documents\Barney Bot\barney_test_microphone_detection.py: 1
# C:\Users\andre\Documents\Barney Bot\barney_test_voice.py: 1
# C:\Users\andre\Documents\Barney Bot\barney_test_voice_andrei_edition.py: 1
# C:\Users\andre\Documents\Barney Bot\jarod-bot.py: 1
# C:\Users\andre\Documents\Barney Bot\members and guild print.py: 3
# C:\Users\andre\Documents\Barney Bot\zinger_bot.py: 2,5
discord.py == 1.7.3

# C:\Users\andre\Documents\Barney Bot\barney.py: 31
emoji == 1.4.2

# C:\Users\andre\Documents\Barney Bot\barney.py: 37
fbchat-asyncio == 0.6.21

# C:\Users\andre\Documents\Barney Bot\graph_test.py: 1
ipywidgets == 7.6.3

# C:\Users\andre\Documents\Barney Bot\graph_test.py: 3
matplotlib == 3.4.2

# C:\Users\andre\Documents\Barney Bot\barney.py: 34
# C:\Users\andre\Documents\Barney Bot\graph_test.py: 2
# C:\Users\andre\Documents\Barney Bot\qr_test.py: 2
numpy == 1.20.1

# C:\Users\andre\Documents\Barney Bot\barney.py: 22
pytesseract == 0.3.8

# C:\Users\andre\Documents\Barney Bot\barney.py: 33
# C:\Users\andre\Documents\Barney Bot\qr_test.py: 1
pyzbar == 0.1.8

# C:\Users\andre\Documents\Barney Bot\barney.py: 36
# C:\Users\andre\Documents\Barney Bot\qr_test.py: 4
qrcode == 7.2

# C:\Users\andre\Documents\Barney Bot\audio_process_test.py: 2
# C:\Users\andre\Documents\Barney Bot\barney.py: 23
# C:\Users\andre\Documents\Barney Bot\barney_chess.py: 3
# C:\Users\andre\Documents\Barney Bot\qr_test.py: 6
# C:\Users\andre\Documents\Barney Bot\schoology_folder_download.py: 2
requests == 2.26.0

# C:\Users\andre\Documents\Barney Bot\schoology_folder_download.py: 3
requests-oauthlib == 1.3.0

# C:\Users\andre\Documents\Barney Bot\barney.py: 24
# C:\Users\andre\Documents\Barney Bot\schoology_folder_download.py: 4
# C:\Users\andre\Documents\Barney Bot\schoology_test.py: 1
schoolopy == 0.2.1

# C:\Users\andre\Documents\Barney Bot\barney_test_microphone_detection.py: 5
sounddevice == 0.4.2

# C:\Users\andre\Documents\Barney Bot\barney_test_voice.py: 3
# C:\Users\andre\Documents\Barney Bot\barney_test_voice_andrei_edition.py: 3
speechrecognition == 3.8.1

# C:\Users\andre\Documents\Barney Bot\barney.py: 29
# C:\Users\andre\Documents\Barney Bot\graph_test.py: 5
# C:\Users\andre\Documents\Barney Bot\intsteps.py: 10
# C:\Users\andre\Documents\Barney Bot\random_integral_generator.py: 2
sympy == 1.8
